requirements.txt
numpy
pandas
scikit-learn
seaborn
matplotlib
python-dotenv   
mysql-connector-python
pymysql   

#-e .
